pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel", "pybind11>=2.12"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-beamforming-toolkit"
version = "0.1.0"
description = "Beamforming simulation toolkit with a C++ compute core, Python API, CLI, and Dash UI."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "schmijul" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: C++",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.13",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
ml = [
  "scikit-learn>=1.5",
]
torch = [
  "torch>=2.3",
]
rl = [
  "gymnasium>=0.29",
]
ui = [
  "dash>=2.18",
  "plotly>=5.24",
]
dev = [
  "pytest>=8.3",
]

[project.urls]
Documentation = "https://github.com/schmijul/Adaptive-Beamforming-Toolkit/tree/main/docs"
Source = "https://github.com/schmijul/Adaptive-Beamforming-Toolkit"
Issues = "https://github.com/schmijul/Adaptive-Beamforming-Toolkit/issues"

[project.scripts]
abf = "abf.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["abf", "abf.*", "algorithms", "core", "data", "simulations", "ui", "visualize"]
